# openapi.yaml
openapi: 3.0.3
info:
  title: Outfitapp API
  version: 0.0.1-SNAPSHOT
  description: |
    Backend API for the Outfitapp wardrobe and outfit-recommendation service.

    **Typical client flow**
    1. `POST /api/users` — create a user (or use an existing `userId`).
    2. `POST /api/uploads/batches` — create an upload batch; response includes signed GCS PUT URLs.
    3. Client uploads each photo with `PUT` to the corresponding `uploadUrl` (GCS, not this API),
       using the same `Content-Type` supplied when creating the batch.
    4. `GET /api/uploads/batches/{batchId}` — poll upload status.
    5. `GET /api/uploads/items/{itemId}/extraction` — poll garment-extraction status for an item.
    6. `GET /api/uploads/items/{itemId}/garments` — fetch detected garments + metadata once extraction completes.
    7. `GET /api/users/{userId}/garments` — list the user's wardrobe (optional metadata filters).
    8. `PATCH /api/users/{userId}/garments/{garmentId}` — rename a garment label for the requested language.
    9. `DELETE /api/users/{userId}/garments/{garmentId}` — soft-delete a garment from the wardrobe.
    10. `POST /api/users/{userId}/outfit-recommendations` — get up to 5 outfit suggestions for a
        free-text context. When `hasMore` is true, call again with prior outfits in `excludeOutfits`
        to request another batch.
    11. `POST /api/users/{userId}/buy-advice` — create a buy-advice job (signed PUT URL for a shopping-photo candidate).
...

...
servers:
...
paths:
  /api/users:
    get:
      tags: [Users]
      operationId: getAllUsers
      summary: List all users
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User already exists (idempotent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "201":
          description: User created
          headers:
            Location:
              description: URI of the created user (`/api/users/{id}`). Note there is currently no GET-by-id endpoint.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
...
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    CreateUserRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated user id
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
    UserResponse:
      type: object
      required: [id, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string

    # --- Enums (serialized as string names) ---
...
    ErrorCode:
      type: string
      description: Machine-readable error code from the API
      enum:
        - EMAIL_ALREADY_EXISTS
        - USER_NOT_FOUND
        - UPLOAD_BATCH_NOT_FOUND
        - UPLOAD_ITEM_NOT_FOUND
        - GARMENT_EXTRACTION_NOT_FOUND
        - GARMENT_NOT_FOUND
        - NO_ELIGIBLE_GARMENTS
        - BUY_ADVICE_NOT_FOUND
        - UNSUPPORTED_IMAGE_FORMAT
        - QWEN_RESPONSE_PARSE_ERROR
        - UNSUPPORTED_LANG
        - VALIDATION_FAILED
        - INTERNAL_ERROR